test5

+++ include/task1.h
@@ -1,6 +1,6 @@
 #ifndef Task1_H
 #define Task1_H
-const size_t LARGE_SIZE = 100000;
+const unsigned int LARGE_SIZE = 100000;
 struct rnd {
     int operator()() {
         return rand() % LARGE_SIZE;
